+++ examples/append.rs
@@ -50,7 +50,7 @@ fn main() {
     }
 
     let params = ImapLoginParams::new(user, SecretString::from(pass)).unwrap();
-    let mut coroutine = ImapLogin::new(params, true);
+    let mut coroutine = ImapLogin::new(params, true, None);
     let mut arg: Option<&[u8]> = None;
 
     loop {